+++ src/java/com/jaeksoft/searchlib/result/collector/JoinDocInterface.java
@@ -26,10 +26,10 @@ package com.jaeksoft.searchlib.result.collector;
 
 public interface JoinDocInterface extends DocIdInterface {
 
-	void setForeignDocId(int pos, int joinResultPos, int foreignDocId);
+	void setForeignDocId(int pos, int joinResultPos, long foreignDocId);
 
-	int getForeignDocIds(int pos, int joinPosition);
+	long getForeignDocIds(int pos, int joinPosition);
 
-	int[][] getForeignDocIdsArray();
+	long[][] getForeignDocIdsArray();
 
 }
