+++ src/java/com/jaeksoft/searchlib/result/collector/MaxScoreCollector.java
@@ -31,7 +31,7 @@ public class MaxScoreCollector extends AbstractCollector {
 	private float maxScore = 0;
 
 	@Override
-	final public void collect(int docId) throws IOException {
+	final public void collect(long docId) throws IOException {
 		float sc = scorer.score();
 		if (sc > maxScore)
 			maxScore = sc;
